[refactor/#62]:phone -> contact 명칭 변경
