+++ src/components/Form/AuthForm.tsx
@@ -11,6 +11,7 @@ function AuthForm() {
     register,
     watch,
     setError,
+
     formState: { errors },
   } = useFormContext();
   const { handleSendCode, expTime, handleVerifyCode } = useCode();
@@ -23,12 +24,12 @@ function AuthForm() {
           type='tel'
           disabled={isVerified}
           placeholder='전화번호를 입력해주세요'
-          {...register('phone', { required: true, pattern: REGEXP.phone })}
+          {...register('contact', { required: true, pattern: REGEXP.contact })}
         />
         <Button
           type='button'
-          disabled={!watch('phone')}
-          onClick={() => handleSendCode(watch('phone'))}
+          disabled={!watch('contact')}
+          onClick={() => handleSendCode(watch('contact'))}
           $style='outlined'
           $size='100px'
           $theme='primary'
@@ -70,7 +71,7 @@ function AuthForm() {
             onClick={
               isVerified
                 ? undefined
-                : () => handleVerifyCode(watch('phone'), watch('code'))
+                : () => handleVerifyCode(watch('contact'), watch('code'))
             }
             $style={isVerified ? 'solid' : 'outlined'}
             $size='100px'
