+++ src/features/Catalog/components/Main/ProductGrid.tsx
@@ -68,8 +68,4 @@ const List = styled.div`
   @media (min-width: 768px) {
     justify-content: space-around;
   }
-
-  @media (min-width: 1024px) {
-    justify-content: space-between;
-  }
 `;
